テストコードのすぐ横でモックする
